requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
scikit-optimize==0.9.0
scipy
numpy
flask
gunicorn
flasgger
Flask-JWT-Extended
python-dotenv